fix: move constants file to helpers folder
